move scalarUtils
