+++ frontend/src/metabase/visualizations/lib/scalarUtils.js
